BUG/MEDIUM: stconn: Really report blocked send if sends are blocked by an error
A regression was introduced in commit 5ec25102ed ("BUG/MEDIUM: stconn:
Report blocked send if sends are blocked by an error") when the patch was
backported. Instead of checking if some outgoing data were blocked, the
opposite was performed. The write timeout must be armed is the channel IS
NOT empty.

It should fix the issue #2754 for the 2.8. It is 2.8-specific. there is no
upstream commit ID and no backport is needed.