BUG/MINOR: spoe: Allow applet creation when closing the last one during stopping
When the last SPOE applet is closed, for any reason, if there are still some
pending messages, in the shared sending queueu or the shared waiting queue, a
new applet is automatically created to be able to pocess these messages. Howver,
this was not performed when HAProxy was stopping. But it was an error because it
was then possible to abort some messages and reported them on error with the
status code set to 256.

So, now, it is still possible to create applet, if necessary, when HAProxy is
stopping.

This patch should partially fix the issue #2868. It must be backported to all
stable versions.

(cherry picked from commit e2ceec1309451f8c304f2fdcef4ced828d57f409)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 26e7ab0ebd8dc4642f2707edf047c774945bac14)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 87e87b810671972bde8210f798bdd7ae467e3af7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>